+++ app/models/EwbsActionRevision.php
@@ -15,6 +15,7 @@ class EwbsActionRevision extends RevisionModel {
 	public static $STATE_TODO='todo';
 	public static $STATE_PROGRESS='progress';
 	public static $STATE_DONE='done';
+	public static $STATE_STANDBY='standby';
 	public static $STATE_GIVENUP='givenup';
 	
 	public static $PRIORITY_LOW='low';
@@ -44,7 +45,7 @@ class EwbsActionRevision extends RevisionModel {
 	 * return array
 	 */
 	public static function states() {
-		return [self::$STATE_TODO, self::$STATE_PROGRESS, self::$STATE_DONE, self::$STATE_GIVENUP];
+		return [self::$STATE_TODO, self::$STATE_PROGRESS, self::$STATE_DONE, self::$STATE_STANDBY, self::$STATE_GIVENUP];
 	}
 	
 	/**
@@ -58,6 +59,7 @@ class EwbsActionRevision extends RevisionModel {
 			case self::$STATE_TODO     :return 'default';
 			case self::$STATE_PROGRESS :return 'primary';
 			case self::$STATE_DONE     :return 'success';
+			case self::$STATE_STANDBY  :return 'info';
 			case self::$STATE_GIVENUP  :return 'warning';
 			default : throw new \UnexpectedValueException($state);
 		}
@@ -74,7 +76,8 @@ class EwbsActionRevision extends RevisionModel {
 			case self::$STATE_TODO     :return 1;
 			case self::$STATE_PROGRESS :return 2;
 			case self::$STATE_DONE     :return 3;
-			case self::$STATE_GIVENUP  :return 4;
+			case self::$STATE_STANDBY  :return 4;
+			case self::$STATE_GIVENUP  :return 5;
 			default : throw new \UnexpectedValueException($state);
 		}
 	}
@@ -195,6 +198,7 @@ class EwbsActionRevision extends RevisionModel {
 				DB::raw("COUNT(DISTINCT CASE WHEN v_lastrevisionewbsaction.state = '".EwbsActionRevision::$STATE_TODO."'     THEN v_lastrevisionewbsaction.id ELSE NULL END) AS count_state_todo"),
 				DB::raw("COUNT(DISTINCT CASE WHEN v_lastrevisionewbsaction.state = '".EwbsActionRevision::$STATE_PROGRESS."' THEN v_lastrevisionewbsaction.id ELSE NULL END) AS count_state_progress"),
 				DB::raw("COUNT(DISTINCT CASE WHEN v_lastrevisionewbsaction.state = '".EwbsActionRevision::$STATE_DONE."'     THEN v_lastrevisionewbsaction.id ELSE NULL END) AS count_state_done"),
+				DB::raw("COUNT(DISTINCT CASE WHEN v_lastrevisionewbsaction.state = '".EwbsActionRevision::$STATE_STANDBY."'  THEN v_lastrevisionewbsaction.id ELSE NULL END) AS count_state_standby"),
 				DB::raw("COUNT(DISTINCT CASE WHEN v_lastrevisionewbsaction.state = '".EwbsActionRevision::$STATE_GIVENUP."'  THEN v_lastrevisionewbsaction.id ELSE NULL END) AS count_state_givenup")
 			]);
 			if(empty($substates)) return;